+++ components/domain/index.js
@@ -8,3 +8,4 @@ export { default as NoticeCard } from './NoticeCard'
 export { default as DashboardCard } from './DashboardCard'
 export { default as ScoreSlider } from './ScoreSlider'
 export { default as EditAboutMe } from './EditAboutMe'
+export { default as InputItem } from './InputItem'
